project_sky_minimal/sky_health_check/requirements.txt
Django==4.2.7
Pillow==10.1.0
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.58.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
Faker==37.1.0

